Rename "order preparation" into "drink preparation"
