+++ README.md
@@ -15,12 +15,12 @@ it checks that the necessary ingredients are available, sends the order for prep
 
 * The current service is in charge of resolving the price and the list of necessary ingredients according to a catalogue.
 * It checks that the ingredients are available by calling another service which exposes current stock status.
-* Orders are sent asynchronously to another service, dedicated to order preparation. 
+* Orders are sent asynchronously to another service, dedicated to drink preparation. 
 
 ![Adapters](./doc/img/Adapters.png)
 
 This translates into 4 ports:
-* A driving port, Order Controller, is in charge of receiving orders.
+* A driving port, Customer Order Handler, is in charge of processing orders.
 * A driven port, Catalogue, exposes the recipes along with their price.
 * A driven port, Stock, exposes current stock status by calling the dedicated service.
-* A driven port, Order Preparation, sends the order to the service in charge of order preparation.
+* A driven port, Drink Preparation, sends the order to the service in charge of drink preparation.
