+++ src/main/java/com/pse/coffee/infra/SpringContextConfiguration.java
@@ -1,8 +1,10 @@
 package com.pse.coffee.infra;
 
-import com.pse.coffee.domain.*;
+import com.pse.coffee.domain.CustomerOrderHandler;
+import com.pse.coffee.domain.OrderingService;
+import com.pse.coffee.domain.Stock;
 import com.pse.coffee.domain.catalogue.Catalogue;
-import com.pse.coffee.domain.preparation.OrderPreparation;
+import com.pse.coffee.domain.preparation.DrinkPreparation;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -12,8 +14,8 @@ import org.springframework.context.annotation.Configuration;
 public class SpringContextConfiguration {
 
     @Bean
-    public CustomerOrderHandler customerOrderHandler(final OrderPreparation orderHandler,
-                                         final Stock stock, final Catalogue catalogue) {
-        return new OrderingService(orderHandler, stock, catalogue);
+    public CustomerOrderHandler customerOrderHandler(final DrinkPreparation drinkPreparation,
+                                                     final Stock stock, final Catalogue catalogue) {
+        return new OrderingService(drinkPreparation, stock, catalogue);
     }
 }
