+++ src/main/java/com/pse/coffee/infra/driven/preparation/DrinkPreparationAdapter.java
@@ -1,7 +1,7 @@
 package com.pse.coffee.infra.driven.preparation;
 
 import com.pse.coffee.domain.preparation.Drink;
-import com.pse.coffee.domain.preparation.OrderPreparation;
+import com.pse.coffee.domain.preparation.DrinkPreparation;
 import com.pse.commons.HexagonalArchitecture;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 @Slf4j
 @Service
 @HexagonalArchitecture.RightAdapter
-public final class OrderPreparationAdapter implements OrderPreparation {
+public final class DrinkPreparationAdapter implements DrinkPreparation {
     @Override
     public void prepare(@NonNull final Drink drink) {
         log.info("Right Adapter: Prepare drink: {}", drink);
